hoprd_api_client

Trait ClientInfo

pub trait ClientInfo<Inner> {
    // Required methods
    fn api_version() -> &'static str;
    fn baseurl(&self) -> &str;
    fn client(&self) -> &Client;
    fn inner(&self) -> &Inner;
}

Interface for which an implementation is generated for all clients.

Required Methods

fn api_version() -> &'static str

Get the version of this API.

This string is pulled directly from the source OpenAPI document and may be in any format the API selects.

fn baseurl(&self) -> &str

Get the base URL to which requests are made.

fn client(&self) -> &Client

Get the internal reqwest::Client used to make requests.

fn inner(&self) -> &Inner

Get the inner value of type T if one is specified.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.