hoprd_api_client

Module types

Types used as operation parameters and responses.

Modules

error

Error types.

Structs

AccountAddressesResponse

Contains the node’s native addresses.

AccountBalancesResponse

Contains all node’s and safe’s related balances.

AnnouncedPeer

Represents a peer that has been announced on-chain.

ApiError

Standardized error response for the API

ChannelInfoResponse

General information about a channel state.

ChannelTicket

Represents a ticket in a channel.

ChannelsQueryRequest

Parameters for enumerating channels.

CloseChannelResponse

Status of the channel after a close operation.

EntryNode

Reachable entry node information

FundBodyRequest

Specifies the amount of HOPR tokens to fund a channel with.

FundChannelResponse

Response body for funding a channel.

HeartbeatInfo

Heartbeat information for a peer.

NodeChannel

Channel information as seen by the node.

NodeChannelsResponse

Listing of channels.

NodeInfoResponse

Information about the current node. Covers network, addresses, eligibility, connectivity status, contracts addresses and indexer state.

NodePeerInfoResponse

Contains the multiaddresses of peers that are announced on-chain and observed by the node.

NodePeersQueryRequest

Quality information for a peer.

NodePeersResponse

All connected and announced peers.

NodeTicketStatisticsResponse

Received tickets statistics.

NodeVersionResponse

Running node version.

OpenChannelBodyRequest

Request body for opening a channel.

OpenChannelResponse

Response body for opening a channel.

PeerObservations

All information about a known peer.

PeerPacketStatsResponse

Packet statistics for a peer.

PingResponse

Contains the latency and the reported version of a peer that has been pinged.

SessionClientRequest

Request body for creating a new client session.

SessionClientResponse

Response body for creating a new client session.

SessionConfig

SessionConfig

SessionStatsAck

Session acknowledgement metrics.

SessionStatsFrameBuffer

Session frame buffer metrics.

SessionStatsLifetime

Session lifetime metrics.

SessionStatsResponse

Complete snapshot of session metrics.

SessionStatsSurb

Session SURB (Single Use Reply Block) metrics.

SessionStatsTransport

Session transport-level metrics.

TicketPriceResponse

Contains the ticket price in HOPR tokens.

TicketProbabilityResponse

Contains the winning probability of a ticket.

WithdrawBodyRequest

Request body for the withdrawal endpoint.

WithdrawResponse

Response body for the withdrawal endpoint.

Enums

IpProtocol

IP transport protocol

RoutingOptions

Request parameters for creating a websocket session.

SessionCapability

Session capabilities that can be negotiated with the target peer.

SessionStatsAckMode

Session acknowledgement mode for metrics.

SessionStatsState

Session lifecycle state for metrics.

SessionTargetSpec

Session target specification.