hoprd_api_client::types

Struct SessionClientRequest

pub struct SessionClientRequest {
    pub capabilities: Option<Vec<SessionCapability>>,
    pub destination: String,
    pub forward_path: RoutingOptions,
    pub listen_host: Option<String>,
    pub max_client_sessions: Option<u64>,
    pub max_surb_upstream: Option<String>,
    pub response_buffer: Option<String>,
    pub return_path: RoutingOptions,
    pub session_pool: Option<u64>,
    pub target: SessionTargetSpec,
}

Request body for creating a new client session.

JSON schema

{
 "description": "Request body for creating a new client session.",
 "examples": [
   {
     "capabilities": [
       "Retransmission",
       "Segmentation"
     ],
     "destination": "0x1B482420Afa04aeC1Ef0e4a00C18451E84466c75",
     "forwardPath": {
       "Hops": 1
     },
     "listenHost": "127.0.0.1:10000",
     "maxClientSessions": 2,
     "maxSurbUpstream": "2000 kb/s",
     "responseBuffer": "2 MB",
     "returnPath": {
       "Hops": 1
     },
     "sessionPool": 0,
     "target": {
       "Plain": "localhost:8080"
     }
   }
 ],
 "type": "object",
 "required": [
   "destination",
   "forwardPath",
   "returnPath",
   "target"
 ],
 "properties": {
   "capabilities": {
     "description": "Capabilities for the Session protocol.\n\nDefaults to `Segmentation` and `Retransmission` for TCP and nothing for UDP.",
     "type": [
       "array",
       "null"
     ],
     "items": {
       "$ref": "#/components/schemas/SessionCapability"
     }
   },
   "destination": {
     "description": "Address of the Exit node.",
     "type": "string"
   },
   "forwardPath": {
     "$ref": "#/components/schemas/RoutingOptions"
   },
   "listenHost": {
     "description": "Listen host (`ip:port`) for the Session socket at the Entry node.\n\nSupports also partial specification (only `ip` or only `:port`) with the\nrespective part replaced by the node's configured default.",
     "type": [
       "string",
       "null"
     ]
   },
   "maxClientSessions": {
     "description": "The maximum number of client sessions that the listener can spawn.\n\nThis currently applies only to the TCP sessions, as UDP sessions cannot\nhandle multiple clients (and spawn therefore always only a single session).\n\nIf this value is smaller than the value specified in `session_pool`, it will\nbe set to that value.\n\nThe default value is 5.",
     "type": [
       "integer",
       "null"
     ],
     "minimum": 0.0
   },
   "maxSurbUpstream": {
     "description": "The maximum throughput at which artificial SURBs might be generated and sent\nto the recipient of the Session.\n\nOn Sessions that rarely send data but receive a lot (= Exit node has high SURB consumption),\nthis should roughly match the maximum retrieval throughput.\n\nAll syntaxes like \"2 MBps\", \"1.2Mbps\", \"300 kb/s\", \"1.23 Mb/s\" are supported.",
     "type": [
       "string",
       "null"
     ]
   },
   "responseBuffer": {
     "description": "The amount of response data the Session counterparty can deliver back to us,\nwithout us sending any SURBs to them.\n\nIn other words, this size is recalculated to a number of SURBs delivered\nto the counterparty upfront and then maintained.\nThe maintenance is dynamic, based on the number of responses we receive.\n\nAll syntaxes like \"2 MB\", \"128 kiB\", \"3MiB\" are supported. The value must be\nat least the size of 2 Session packet payloads.",
     "type": [
       "string",
       "null"
     ]
   },
   "returnPath": {
     "$ref": "#/components/schemas/RoutingOptions"
   },
   "sessionPool": {
     "description": "How many Sessions to pool for clients.\n\nIf no sessions are pooled, they will be opened ad-hoc when a client connects.\nIt has no effect on UDP sessions in the current implementation.\n\nCurrently, the maximum value is 5.",
     "type": [
       "integer",
       "null"
     ],
     "minimum": 0.0
   },
   "target": {
     "$ref": "#/components/schemas/SessionTargetSpec"
   }
 }
}

Fields
capabilities: Option<Vec<SessionCapability>>

Capabilities for the Session protocol.

Defaults to Segmentation and Retransmission for TCP and nothing for UDP.

destination: String

Address of the Exit node.

forward_path: RoutingOptionslisten_host: Option<String>

Listen host (ip:port) for the Session socket at the Entry node.

Supports also partial specification (only ip or only :port) with the respective part replaced by the nodeโs configured default.

max_client_sessions: Option<u64>

The maximum number of client sessions that the listener can spawn.

This currently applies only to the TCP sessions, as UDP sessions cannot handle multiple clients (and spawn therefore always only a single session).

If this value is smaller than the value specified in session_pool, it will be set to that value.

The default value is 5.

max_surb_upstream: Option<String>

The maximum throughput at which artificial SURBs might be generated and sent to the recipient of the Session.

On Sessions that rarely send data but receive a lot (= Exit node has high SURB consumption), this should roughly match the maximum retrieval throughput.

All syntaxes like โ2 MBpsโ, โ1.2Mbpsโ, โ300 kb/sโ, โ1.23 Mb/sโ are supported.

response_buffer: Option<String>

The amount of response data the Session counterparty can deliver back to us, without us sending any SURBs to them.

In other words, this size is recalculated to a number of SURBs delivered to the counterparty upfront and then maintained. The maintenance is dynamic, based on the number of responses we receive.

All syntaxes like โ2 MBโ, โ128 kiBโ, โ3MiBโ are supported. The value must be at least the size of 2 Session packet payloads.

return_path: RoutingOptionssession_pool: Option<u64>

How many Sessions to pool for clients.

If no sessions are pooled, they will be opened ad-hoc when a client connects. It has no effect on UDP sessions in the current implementation.

Currently, the maximum value is 5.

target: SessionTargetSpec